requirements/base.txt
Django==2.0.5
psycopg2==2.7.3.2
django-crispy-forms==1.7.0
django-model-utils==3.1.1
python-dateutil==2.6.1
django-filter==1.1.0
django-allauth==0.35.0
Pillow==5.1.0
